Update Node base image and improve build process
This commit updates the Node base image to 20.11.1-alpine3.18 in multiple build scripts. Additionally, it improves the build process by adding build status messages. Now, if a build fails, it will print the contents of the source package and certain .js files to facilitate debugging, and the findutils package was added to the Docker image.